feat: add onRaceClick logic

+++ src/ts/actions/garageActions.ts
@@ -1,5 +1,5 @@
 import API from '../api'
-import { startDriving, stopDriving, updateStateGarage } from '../controllers'
+import { startDriving, stopDriving, updateStateGarage, updateStateWinners } from '../controllers'
 import { CarI, GarageTotalTargets, StoreI } from '../interfaces'
 import { carList } from '../templates/carList'
 import utils from '../utils'
@@ -17,11 +17,12 @@ class GarageActions {
         this.addListenersOnCarItem()
 
         const { generatorBtn, raceBtn, resetBtn } = utils.findRaceControlsButtons()
-        generatorBtn.addEventListener('click', (e) => this.onGeneratorClick(e, this.store))
+        generatorBtn.addEventListener('click', (e) => this.onGeneratorClick(e))
+        raceBtn.addEventListener('click', (e) => this.onRaceClick(e))
     }
 
     private addListenersOnCarItem = () => {
-        const { startBtnList, stopBtnList, selectCarBtnList, deleteCarBtnList } =
+        const [startBtnList, stopBtnList, selectCarBtnList, deleteCarBtnList] =
             utils.findCarListsButtons()
 
         startBtnList.forEach((el) => {
@@ -34,7 +35,7 @@ class GarageActions {
             el.addEventListener('click', this.onSelectClick)
         })
         deleteCarBtnList.forEach((el) => {
-            el.addEventListener('click', (e) => this.onDeleteClick(e, this.store))
+            el.addEventListener('click', (e) => this.onDeleteClick(e))
         })
     }
 
@@ -66,46 +67,90 @@ class GarageActions {
         }
     }
 
-    private async onDeleteClick(e: MouseEvent, store: StoreI) {
+    private async onDeleteClick(e: MouseEvent) {
         const targetBtn = <HTMLElement>e.target
         targetBtn.textContent = `Loading...`
         const id = +targetBtn.id.split('remove-car-')[1]
-        const carEl = <HTMLElement>document.getElementById(`car-item${id}`)
         const totalCountEl = <HTMLElement>(
             document.getElementById(`${GarageTotalTargets.TotalCount}`)
         )
         const carsListEl = <HTMLElement>document.getElementById(`${GarageTotalTargets.GarageList}`)
 
         await API.deleteCar(id)
-        await API.deleteWinner(id)
-        await updateStateGarage(store)
+        const isCarInWinners = this.store.winners.find((el) => el.id === id)
+        if (isCarInWinners) {
+            await API.deleteWinner(id)
+        }
 
-        totalCountEl.textContent = `Garage (${store.carsCount} cars)`
-        carsListEl.innerHTML = carList(store.cars)
+        await updateStateGarage(this.store)
+
+        totalCountEl.textContent = `Garage (${this.store.carsCount} cars)`
+        carsListEl.innerHTML = carList(this.store.cars)
         this.addListenersOnCarItem()
     }
 
-    private async onGeneratorClick(e: MouseEvent, store: StoreI) {
+    private async onGeneratorClick(e: MouseEvent) {
         const targetBtn = <HTMLButtonElement>e.target
         const totalCountEl = <HTMLElement>(
             document.getElementById(`${GarageTotalTargets.TotalCount}`)
         )
+        const carsListEl = <HTMLElement>document.getElementById(`${GarageTotalTargets.GarageList}`)
 
         targetBtn.textContent = `Generating cars...`
         targetBtn.disabled = true
 
         const cars = utils.generateRandomCars()
-
         await Promise.all(cars.map(async (car) => await API.createCar(car)))
-        await updateStateGarage(store)
-        const { items, totalCount } = await API.getCars(1)
-
-        this.store.cars = items
-        this.store.carsCount = totalCount
+        await updateStateGarage(this.store)
 
         targetBtn.textContent = `Generate cars`
         targetBtn.disabled = false
-        totalCountEl.textContent = `Garage (${store.carsCount} cars)`
+        totalCountEl.textContent = `Garage (${this.store.carsCount} cars)`
+        carsListEl.innerHTML = carList(this.store.cars)
+        this.addListenersOnCarItem()
+    }
+
+    private async onRaceClick(e: MouseEvent) {
+        const messageEl = <HTMLElement>document.getElementById(GarageTotalTargets.Message)
+        const targetBtn = <HTMLButtonElement>e.target
+        const carsListEl = <HTMLElement>document.getElementById(`${GarageTotalTargets.GarageList}`)
+
+        targetBtn.disabled = true
+        this.store.isRace = true
+
+        utils.findCarListsButtons().forEach((list) => {
+            list.forEach((el) => {
+                el.disabled = true
+            })
+        })
+        utils.findSubmitButtons().forEach((el) => {
+            el.disabled = true
+        })
+
+        const promises = this.store.cars.map((car: CarI) => startDriving(car.id))
+        const responses = await Promise.all(promises)
+        const winner = responses
+            .filter((el) => el.status === 200)
+            .sort((a, b) => a.time - b.time)[0]
+        await API.saveWinner(winner.id, winner.time)
+        const winnerCar = await API.getCar(winner.id)
+        await updateStateWinners(this.store)
+
+        messageEl.innerHTML = `${winnerCar.name} went first (${utils.getTimeInSeconds(
+            winner.time
+        )}s)!`
+        messageEl.classList.add('visible')
+        utils.findSubmitButtons().forEach((el) => {
+            el.disabled = false
+        })
+        targetBtn.disabled = false
+        this.store.isRace = false
+
+        setTimeout(() => {
+            messageEl.classList.remove('visible')
+            carsListEl.innerHTML = carList(this.store.cars)
+            this.addListenersOnCarItem()
+        }, 5000)
     }
 }
